Project Web Application Developing a pressure shoe for saveing data walking
Health is an important basis for the development of a person's quality of life. Health refers to a state of complete physical, mental, and social well-being. According to a survey of health conditions of Thai people across the country. The overall physical health of Thai people in all areas tends to improve if the average life expectancy at birth tends to increase in all regions. especially in the west, the south and Bangkok from entering the aging society. mortality and disability rates per population in every region rising trend especially in the North and Bangkok with the highest mortality rates of 8.5 and 8.2 deaths per 1,000 population. For causes of death and disability in 2019, it was found that the top three causes were cancer and all tumor types. Circulatory and respiratory diseases, both men and women in Bangkok have higher rates of mortality and disability. from cancer and tumors and high respiratory disease compared to other sectors For men in the northern region, the death rate from respiratory disease was highest in all regions. Women in the North had a higher cause of death from circulatory diseases than women in other regions. (Thai Health Report, 2021)